requirements.txt
streamlit==1.32.0
groq==0.4.2
google-generativeai==0.4.1
python-docx==1.1.0
PyMuPDF==1.23.26
requests==2.31.0
python-dotenv==1.0.1
